eve: Add 2 minute timer and double tap for both LEDs
To match the final LED spec, the low battery patterns should only be
displayed for 2 minutes whenever a low battery pattern threshold is
crossed. In addition, when double tap is active (not in S0), the
double tap pattern should be displayed on both LEDs. For the charging
LED, the requires a transition to OFF, then pattern, transition to
OFF, and back to the charging pattern.

BUG=b:35584895
BRANCH=eve
TEST=Manual Used battfake command to force different battery charge
levels. Connectecd/disconnected external charger. Verified that low
battery patterns display for fixed time, but are preempted when the
charger is connected, except for BLINK_RED pattern on non-charging
LED. Tested various combinations of double tap and
connecting/disconnecting external charger.

Change-Id: Idb36356b5df47ec5346ceea0ef968979c7a26080
Signed-off-by: Scott Collyer <scollyer@google.com>