Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> + smbarber, who wrote this :)
> 
> > we'll see the host asks for alarm set in something far different from timeout. Maybe AP thinks its time and ec etc clock is not in sync?
> 
> https://chromium.googlesource.com/chromiumos/third_party/kernel/+/chromeos-4.4/drivers/rtc/rtc-cros-ec.c#173
> 
> The kernel does:
> 
> 1. read RTC
> 2. calculate offset (since user space gave an absolute alarm time)
> 3. sanity check
> 4. program the offset
> 
> Unfortunately, step 3 looks like:
> 
>                 /* Don't set an alarm in the past. */
>                 if ((u32)alarm_time < current_time)
>                         alarm_offset = EC_RTC_ALARM_CLEAR;
>                 else
>                         alarm_offset = (u32)alarm_time - current_time;
> 
> So, if the EC is too fast, we'd hit the first branch (doesn't look too good). If the EC is too slow, we hit the latter case, and 'alarm_offset' will be too big (diff'ing the local "alarm_time" which will be accurate, with the EC's "current_time" which is not).
> 
> I wonder if we should try to correct this partially in the kernel driver (e.g., use the system time as "current_time", not the RTC time).

I'd be okay with this, and we get to eliminate another host command to boot. There really isn't a change in functionality here since the EC only gets alarms as an offset from the current time anyway.

When I wrote this initially I assumed the EC would be at least as good as the SoC for timekeeping; if the EC is going to drift this much then we should probably prefer the system time as the source of truth :\

Patch-set: 1