Update patch set 1
Patch Set 1:

> > I wonder if we should try to correct this partially in the kernel driver (e.g., use the system time as "current_time", not the RTC time).

Actually I was wrong about this; the kernel RTC framework is internally consistent, by always dealing with the RTC time (not system time). So rtc-cros-ec is more or less correct.

The problem is that *powerd* is not consistent. It's using the system time in powerd_dbus_suspend, instead of just using the relative offset form (e.g., "+XXX") to let the kernel do RTC-relative computations:

https://chromium.googlesource.com/chromiumos/platform2/+/master/power_manager/tools/powerd_dbus_suspend.cc#92

I'll patch that.

> I'd be okay with this, and we get to eliminate another host command to boot. There really isn't a change in functionality here since the EC only gets alarms as an offset from the current time anyway.
> 
> When I wrote this initially I assumed the EC would be at least as good as the SoC for timekeeping; if the EC is going to drift this much then we should probably prefer the system time as the source of truth :\

I'm not sure which one we should be expect to be more accurate. The local CPU provides a lot more finer-grained timekeeping at least. Anyway, we definitely need to improve the EC accuracy, but we should also fix up our tools so that they don't rely on the RTC and the AP being completely in sync.

Patch-set: 1
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>