Update patch set 1
Patch Set 1:

The issue is that this only applies to a single battery. I'm not sure of the implications for the other batteries that coral is utilizing. Can we enable this only take action when we indicate that the battery needs this?

Also, where is the bug for this?

Patch-set: 1
Reviewer: Gerrit User 1002124 <1002124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>