Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I see you marked this CL private. Are you sure you want to disclose these unseal codes?

Yeah that was the concern so we wanted to keep the commit private..

New batteries are going to contain this change, but we have several thousand that are already manufactured and going into systems so we're trying to figure out the best way to roll it out.

One thought was to have an Eve EC release allow tunneling i2c to the battery so we could do it in a utility that lives in the private overlay.

Patch-set: 1
Reviewer: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>