Update patch set 1
Patch Set 1:

(3 comments)

I changed this so that it's all done in terms of i2cxfer, instead of calls to sb_write or sb_read_mfgacc so that it will be easier to convert to a script using ectool i2cxfer.

I also added a time delay if the flash area is written. I found with testing that the seal would often fail when the contents of 0x482c is changed. With delay, it was successful each time.

Patch-set: 1
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>