Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> Is there a non-interactive version of this as well?
> 
> The recovery shim will get the challenge and format it as a QR code and display it via frecon.  It will get the response via frecon and then send it back.

I was thinking this could be used in non-interactive mode just as well, the utility would be called, the challgenge retrieved, the process which started the utility would keep running waiting for input, once the auth code is available it is sent to the Cr50.

Will this not work?

Patch-set: 4
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>