cr50: Add uartblock command to block enabling EC/AP UARTs
Currently, when CCD is opened, there is no way to disable the EC
and/or AP UARTs.  But if there is some problem with the EC and/or AP,
and their UARTs are spamming interrupts, it can make debugging more
difficult.

Add a new uartblock command to do this until the next cr50 reboot.

BUG=b:65639347
BRANCH=cr50
TEST=manual with CR50_DEV=1 image, AP/EC powered on, suzyq connected
	uartblock --> (none)
	ccdstate --> UARTAP+TX UARTEC+TX I2C SPI

	uartblock AP on
	ccdstate --> UARTEC+TX I2C SPI

	uartblock EC on
	ccdstate --> I2C SPI

	uartblock -> AP EC

	uartblock AP off
	ccdstate --> UARTAP+TX I2C SPI

	uartblock EC off --> (none)
	ccdstate --> UARTAP+TX UARTEC+TX I2C SPI

	ccd lock
	uartblock AP on --> access denied

Change-Id: I3dcc8314fc98a17af57f2fe0d150ecd1a19ccf52
Signed-off-by: Randall Spangler <rspangler@chromium.org>