Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

the brand code string is stored in the Cr50 as a 4 byte number and its bytes are swapped so that when it is displayed as a 4 byte number, the order of bytes in the number match the order of letters in the string.

So 'ZZCR' which would have been displayed as 0x52435a5a is converted into 0x5a5a4352 before programming into the board ID field, and then is presented as a 4 byte number everywhere.

The server code should just swap the 4 bytes of the RLZ to follow suite.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Trybot-Ready=0