NVMEM: Add Non-volatile memory
CrOS EC currently has access to a few non-volatile memories but
they're not consistently available:

BBRAM: Requires power. Data is lost when power is removed by battery
       cutoff, removal, or hibernation on z-state system.
EEPROM: Write-protect pin affects the entire region. Maskrom change
        is needed to do half-region write-protect.

NVMEM is a data system stored on a flash memory. It should behave
more or less the same on all platforms.

A drawback is that the write latency varies. So, it doesn't suit
for application requiring precise timing.

BUG=b:70294260
BRANCH=none
TEST=buildall

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I3c3eaa3995bb088d88c1891a312d84f974b228a7