Update patch set 3
Patch Set 3:

(10 comments)

It still would be great to have some doc or a buganizer entry to describe what's the intended use of this facility.

It feels like it was extensively discussed over the email or in person, but little record is there describing the intentions and use of this.

a few more comments and nits.

Patch-set: 3
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>