eve: Make BATTERY_DISCONNECTED depend only on XDSG
The battery_check_disconnect() function was checking for both XCGH and
XDSG to be set before returning BATTERY_DISCONNECTED. This works
consistently when recovering from a battery cutoff when issued via a
I2C command. However, if the battery cutoff was initiated due to a
cell under voltage (CUV) trip, only the discharge FET will be
disabled. When an external charger is connected, the battery may not
be able to recover if the CUV event happened while SOC > 3%.

By modifying the battery_check_disconnected to look only at XDSG then
the normal battery cutoff and CUV triggered cutoff are recovered
properly when an external charger is connected.

BUG=b:67332823
BRANCH=eve
TEST=Had two different units where I was able to confirm that a CUV
event had happened. I obsereved that the system would brown out when
the AP was powered up, preventing the EC from being able to charge the
battery to get out of the CUV condition. Once I loaded FW that only
checked XDSG, then verified that the AP was not powered on and the
battery was able to be charged by the EC.

Change-Id: I893cfcdd4123810137e2d09a15256d51c918b38b
Signed-off-by: Scott Collyer <scollyer@google.com>