make: rebuild board image if essential make variables change
Some make variables get converted into compilation flags, which affect
image composition. Changes of these variables go unnoticed as they do
not directly affect make dependencies.

Let's define the set of essential variables in
board/${BOARD}/make_vars if required, and save the state of the
variables at build time in a file in the build output directory.

If any of the variables of the set changes during a following build,
remove the entire contents of the output directory, triggering
rebuilding of the image from scratch.

BRANCH=cr50
BUG=none
TEST=verified that changing of CR50_DEV value triggers full rebuild of
     the Cr50 image.

Change-Id: Id0589a3b6a66fe4da90a9aea894bc83eb6337c8c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>