build: tighten rebuild rules
Ensure that we force rebuilding the source files if we have
changed the compilation command-line options by modifying the
makefiles or passing extra variables.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=cr50
BUG=none
TEST=do 'make BOARD=cr50' without and with CR50_DEV=1 and see the .o
files are re-build when we change from one to the other. re-test the
null rebuild without file change, touch one C file and rebuild.

Change-Id: Ibbf4d7a3d44c17b8b99841514fcb81eb2fc9ff3c