Update patch set 1
Patch Set 1:

I will work with nsanders@ today to make sure his test cases still work and we don't cause any major regressions.

I think we should evaluate this patch series as a whole and land it if it is correct + good for the future of the codebase. servo_v4 has some existing problems (and one or two "new" ones which are exposed by this patch series) and I'll help to fix those. But we also have a FW branch for servo_v4, so the impact of such regressions should be minor.

Patch-set: 1
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>