kionix: Issue soft reset before checking whoami result
Issue a soft reset before checking the whoami register against
the expected value.

The early whoami check is left in place to probe for when the chip
is responding on i2c, but the returned value is not checked.

BUG=b:67865186
BRANCH=eve
TEST=manual on eve: 'crash watchdog' and verify that kxcj9
initializes properly and is functional after the EC crash.

Change-Id: I5c4e2510c8e913a9160838923354d9a63e5acbd4
Signed-off-by: Duncan Laurie <dlaurie@google.com>