chg_mgr: Add dynamic input current limit setting.
Depending on the charge inductor's design, the current limit may need to
be reduced at a certain voltage.  Therefore, some new CONFIG_* were
introduced.

  CONFIG_CHARGER_DYNAMIC_ICL
  CONFIG_CHARGER_DYNAMIC_ICL_VOLTAGE_THRESHOLD

The units are mA and mV respectively.  When CONFIG_CHARGER_DYNAMIC_ICL
is defined, when the charger voltage crosses the threshold, the input
current will be limited to CONFIG_CHARGE_DYNAMIC_ICL.

BUG=b:67020411
BRANCH=None
TEST=Enable this on zoombini; Plug in a charger that can exceed the
configured limit; Verify that when the voltage crosses the threshold,
the current limit is limited to CONFIG_CHARGE_DYNAMIC_ICL.

Change-Id: I5343f8aa4a249a39d48cd4156be599492d729aa4
Signed-off-by: Aseda Aboagye <aaboagye@google.com>