Update patch set 1
Patch Set 1:

The only time it might be concerning, which I hinted at before, is if we select the wrong port, eg. port A provides 3A @ 20V, port B provides 2.9A @ 18V. We'd want to select B.. but if we don't make any changes, we'd select A. That's a very uncommon case, and the negative effect of choosing the wrong port is very minor, so I wouldn't worry about it.

Patch-set: 1
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>