zoombini: Raise input voltage ref for BC1.2 chargers.
When we detect a high powered BC 1.2 charger and try to charge from it,
we set the current limit and let the charger ramp via a input voltage
regulation loop.  However, the default voltage reference for this loop
is 4096 mV. In order to be kinder to BC 1.2 chargers, this commit sets
the input voltage limit to ~4.4V when decide to charge from it.  When we
decide from a different supplier, the reference voltage is set back to
its default of 4096mV.

BUG=b:67964166
BRANCH=None
TEST=Flash zoombini; With a drained battery and a DCP charger and verify
that Vbus does not drop below ~4.4V.

Change-Id: Id1b0885e8e7ee8a974f437598632d96c25c9de17
Signed-off-by: Aseda Aboagye <aaboagye@google.com>