power/rk3399: Increase the power initialization timeout
Currently we are setting the timeout to 40 * 100ms, which is not
enough for bob's battery initialization.

The pack vendor said the internal battery wakeup could take 5+ seconds.

So let's increase the timeout to 100 * 100ms for the slow battery wake.

Also rewrite the comment as it's not only about the charger.

BUG=b:68498722
BRANCH=gru
TEST=Manually test cutoff on bob, no more "power-up inhibited"

Change-Id: Ib7e7f4a6ebb592ee38b4a055cc88d9c808fdf179
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>