Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not sure this behavior is desirable, it may lead to 'zombie' devices that power-on at unexpected times.

That is why I have restricted the check to ensure we power on only if the boot-up was prohibited because of charger issues. i.e. power_up_inhibited to set to true only if EC was attempting to boot AP, but could not do so because of charger issues.

If the AP is powered off for any other reason, then this auto power-on should not be triggered. Basically, the issue as described in the bug mentioned in commit message is:
1. When battery enters critical state, EC hibernates. On connecting AC power, we boot the AP up.
2. If the battery SOC is below required minimum, then the behavior is different i.e. we prohibit AP from booting up. However, the device stays off even after the SOC has reached the required minimum. This change intends to power on the AP only when it was inhibited from booting because of battery SOC.

Patch-set: 1
Reviewer: Gerrit User 1003000 <1003000@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>