Update patch set 1
Patch Set 1:

> That is why I have restricted the check to ensure we power on only if the boot-up was prohibited because of charger issues.

I understand, but it's not clear to me that automatically powering on once battery ticks up to a sufficient level is the best course of action. Let's ask other reviewers what they think.

To be clear, the question is -- if the device would power on (due to AC plug / EC cold boot on dead / cutoff battery, or power button press), but the power-on is inhibited (due to critical low / no battery), is it desirable to automatically power-on seconds to minutes later once the battery is no longer critically low?

Patch-set: 1
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>