Update patch set 2
Patch Set 2:

> Shawn, what do you think?

We should think about the exact conditions for which we wish to cut-off the battery (eg. when we're not charging and SOC < 1%, or voltage < XXXX mV) and implement that. "3600 seconds after SOC drops < 3%" is probably not exactly what we want.

Patch-set: 2
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>