zoombini: Set input voltage regulation for BC1.2.
The ISL9238 has a input voltage regulation loop, which will help prevent
the input voltage from falling significantly below a set threshold.
This is the mechanism when CONFIG_CHARGER_RAMP_HW is defined for this
charger, however we are also using software charge ramping.  The
combination of the two will help us charge from chargers which
overcurrent before crossing the minimum voltage threshold and Vbus
collapses rapidly (turned off).

There is a caveat here for chargers which overcurrent sometime after
ramping past their limit but meeting the minimum input voltage.  It may
take a while for the software charge ramping algorithm to find the true
current limit, but it will happen eventually.  This is because we set
the limit to 2400mA due to our limitations with the BQ24392.

BUG=b:67964166
BRANCH=None
TEST=Plug in a BC1.2 charger, verify that the input voltage regulation
loop is enabled.
TEST=Plug in a PD charger, verify that the input voltage regulation loop
is disabled.

Change-Id: I0054f5fc5aa4e4246e191fc294b75657c25e7635
Signed-off-by: Aseda Aboagye <aaboagye@google.com>