Bob: Disable MPU locking
Port changes cl/413584 to bob.

Our RAM / code regions aren't a power of 2, so we cannot program MPU to
precisely protect the regions we desire.

Shawn N opened crbug.com/782244, this issue will be fix soon, but the
changes will not backporting to this branch due to stability concerns.
So just add the #undef to Board bob to avoid the error message.

BUG=chrome-os-partner:57789
BRANCH=gru
TEST=`make buildall -j`, then burn + boot to OS.

Change-Id: Ibdbe494f6d98319c88577345c2c100e6a7946e30
Signed-off-by: james_chao <james_chao@asus.com>