scarlet: battery: Fix charging allowed state
The max17055 fuel gauge used on scarlet (rev1+) doesn't report whether
the battery allows charging or not. On systems with fuel gauges that do
report it (like the bq27541) the fuel gauge driver is in charge of
setting the BATT_FLAG_WANT_CHARGE flag if charging is allowed. On
scarlet the flags needs to be set by the battery driver whenever
charging is not disallowed (e.g. because the battery temperature is
too high/low), otherwise the battery info will always report that
charging is not allowed.

BUG=chromium:67753608
TEST=see related max17055 CL
BRANCH=none

Change-Id: I680b42f13a4f6282a290e59755949f200a0e8551
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>